+++ atom/browser/api/atom_api_web_contents.cc
@@ -18,6 +18,7 @@
 #include "base/strings/string_util.h"
 #include "base/strings/utf_string_conversions.h"
 #include "brightray/browser/inspectable_web_contents.h"
+#include "chrome/browser/printing/print_view_manager_basic.h"
 #include "content/public/browser/favicon_status.h"
 #include "content/public/browser/guest_host.h"
 #include "content/public/browser/navigation_details.h"
@@ -40,6 +41,15 @@
 
 #include "atom/common/node_includes.h"
 
+namespace {
+
+struct PrintSettings {
+  bool silent;
+  bool print_background;
+};
+
+}  // namespace
+
 namespace mate {
 
 template<>
@@ -60,6 +70,19 @@ struct Converter<atom::api::SetSizeParams> {
       out->max_size.reset(new gfx::Size(size));
     if (params.Get("normal", &size))
       out->normal_size.reset(new gfx::Size(size));
+    return true;
+  }
+};
+
+template<>
+struct Converter<PrintSettings> {
+  static bool FromV8(v8::Isolate* isolate, v8::Local<v8::Value> val,
+                     PrintSettings* out) {
+    mate::Dictionary dict;
+    if (!ConvertFromV8(isolate, val, &dict))
+      return false;
+    dict.Get("silent", &(out->silent));
+    dict.Get("printBackground", &(out->print_background));
     return true;
   }
 };
@@ -135,6 +158,7 @@ WebContents::WebContents(const mate::Dictionary& options)
 
   InitWithWebContents(web_contents, owner_window);
   inspectable_web_contents_ = managed_web_contents();
+  printing::PrintViewManagerBasic::CreateForWebContents(web_contents);
 
   Observe(GetWebContents());
 }
@@ -580,6 +604,17 @@ void WebContents::UnregisterServiceWorker(
                                    callback);
 }
 
+void WebContents::Print(mate::Arguments* args) {
+  PrintSettings settings = { false, false };
+  if (args->Length() == 1 && !args->GetNext(&settings)) {
+    args->ThrowError();
+    return;
+  }
+
+  printing::PrintViewManagerBasic::FromWebContents(web_contents())->
+      PrintNow(settings.silent, settings.print_background);
+}
+
 void WebContents::Undo() {
   web_contents()->Undo();
 }
@@ -750,6 +785,7 @@ mate::ObjectTemplateBuilder WebContents::GetObjectTemplateBuilder(
         .SetMethod("unregisterServiceWorker",
                    &WebContents::UnregisterServiceWorker)
         .SetMethod("inspectServiceWorker", &WebContents::InspectServiceWorker)
+        .SetMethod("print", &WebContents::Print)
         .Build());
 
   return mate::ObjectTemplateBuilder(
